move methods that should be accessible by templates into traits
